Data for Genotypes by Environment Interaction (GEI)

Maiz is used for performing Genotypes by Environment Interaction (GEI) Analysis.

Usage

data(Maiz)

Format

A data.frame 1320 obs. of 6 variables.

Details

Gen Genotype

Institute Institute

Rep Replicate

Block Block

Env Environment

Yield Yield Response

References

Crossa, J., Perez-Elizalde, S., Jarquin, D., Cotes, J.M., Viele, K., Liu, G., and Cornelius, P.L. (2011) Bayesian Estimation of the Additive Main Effects and Multiplicative Interaction Model Crop Science, 51, 1458–1469. (doi: 10.2135/cropsci2010.06.0343)

Author

Muhammad Yaseen (myaseen208@gmail.com)

Jose Crossa (j.crossa@cgiar.org)

Sergio Perez-Elizalde (sergiop@colpos.mx)

Diego Jarquin (diego.jarquin@gmail.com)

Jose Miguel Cotes

Kert Viele

Genzhou Liu

Paul L. Cornelius

Examples

data(Maiz)